+++ packages/kit/src/extensions/feature/index.ts
@@ -1,3 +1,2 @@
 export * from "./capability";
-export * from "./FeatureTool";
 export * from "./FeatureCapabilityExtension";