+++ packages/kit/src/extensions/image/index.ts
@@ -1,5 +1,4 @@
 export * from "./capability";
-export * from "./ImageTool";
 export * from "./ImagePlacementCapabilityExtension";
 export * from "./commands";
 export * from "./config";