+++ packages/kit/src/extensions/index.ts
@@ -10,4 +10,3 @@ export * from "./film";
 export * from "./mirror";
 export * from "./ruler";
 export * from "./white-ink";
-export * from "./dieline-workflow";