+++ packages/kit/src/extensions/size/index.ts
@@ -1,3 +1,2 @@
 export * from "./capability";
-export * from "./SizeTool";
 export * from "./SizeCapabilityExtension";