+++ packages/kit/src/extensions/white-ink/index.ts
@@ -1,5 +1,4 @@
 export * from "./capability";
-export * from "./WhiteInkTool";
 export * from "./WhiteInkCapabilityExtension";
 export * from "./commands";
 export * from "./config";