+++ packages/kit/src/index.ts
@@ -1,2 +1,6 @@
 export * from "./extensions";
 export * from "./factories";
+export {
+  KIT_LEGACY_LAYER_PRESET,
+  type KitLegacyLayerPreset,
+} from "./shared/constants/layers";